undeeped uneccesarry nested routes,iterm & main components deleted
